File: read db record back into object after save
File expects MySQL to generate the timestamps for created and updated. However, we never read the newly created mysql values back into the object after save.
This became a problem when we added Solr to the mix.  Indexing in Solr was the first time we attempted to continue using the File objects after calling save().

Updates #304