Fixed: member update form no longer populates end with current date
We have been mistakenly putting the current date into the member update form.  This came from when we started using \DateTime objects as values.  If you create a new \DateTime object with an empty constructor, you get the current date.
In previous code, we always used the unixtime format ('U'), which results in zero if the date is empty.

Updates #154
Updates #289