Organize meeting data by event_id as well as date
This adds another level of organization to the meeting data feed.  Meetings are now organized by date and event_id.  Meetings that are not in Google Calendar will have an event_id of zero.  This means if there are multiple meetings on the same day that are not in a calendar, they will still be lumped together.

We do not have time information in the database for meetingFiles.  However, for meetings that are in Google Calendar, we can display the start time of the event.
Ultimately, we need to do a database schema change to change the meetingFiles.date column to be a datetime.  Then we'll need to run through all the Google Calendar events and populate the datetime information.

Until then, the only place we do not have the time information is on the MeetingFiles/Index controller.  That is purely reading from the database.

Updates #310