Install libreoffice-java-common
We are using libreoffice to conver files to PDF.  This runs as www-data (the
Apache user).  The user needs to be able to write to it's home directory, so the
dconf cache file can be written.

Also, there was a missing java executable (javaldx).  The javaldx doesn't seem
to be strictly necessary; however, it does produce errors in the error log.
Installing the additional java binaries for libreoffice resolves this.

Updates #319