Member resign form does not choose endDate past Term endDate
When choosing the autofill end date for the Member Resign form, the form no longer simply chooses the current date.  Staff doing the data entry can be updating the data long after the Term has ended.  The autofilled date should not simply use the current date, but cap  the end date at the end date of the Term.

Updates #329