Started work on Solr indexing via cron
Indexing a file when it is uploaded can take a while, it stalls the webpage during this process.  Additionally, it could error out completely.

This is some stubbed out code to show what it would look like to keep track of when a file is indexed and provide a script to index the files that have not been indexed.

Updates #316