Suppport both twig and legacy templating
We probably want to rewrite the HTML one controller at a time, while keeping all of the application functional.  The View class needs to support both the old template/blocks system as well as modern Twig.

Each legacy controller action can be rewritten as an invokable controller.  This way, we can keep track of what's been rewritten in the routes file.