ActionLinks template now expects explicit class
This used to be using the index of the array as the class name.  This led to class names that were numbers.  It also had the problem that items in the list could not use the same class name.

Now, the actionLinks arrays expect the classname to be included in the data.
links[] = ['url' => '', 'label'=>'', 'class' => ''];