Choose Person Callback route cleaned
Adding a new person to the system during the [Person Chooser] process relied on the callback route, which had not been ported over to the new URL structure.

This commit brings in the callback route as part of [People].  The only callback route in the system is for the [Person Chooser].
This rewrite requires javascript to function, as it relies on calling a javascript function on the opening page.

The [Person Chooser] is no longer expected to work in a single window or without javascript.