Fixed: committee::seats switch between current and past
The $_GET['current'] parameter needs to be passed along all the way to SeatTable::find().  However, we no longer want to blindly forward whatever junk a user might have put in the URL.

This cleanly checks the $_GET['current'] parameter and forwards a boolean along to all the function calls.