Updated OIDC claims to use Standard Claims
We ran into an issue with another project where the OIDC library was expecting the Standard Claims.  That library did not have the ability to map claims, thus requiring the Standard Claims.

We updated ADFS to give out the Standard Claims, rather than the custom names ADFS was offering.  Even though we're still mapping them here, it simplifies our deplopyments if everyone agrees to the use the same set of claims.