Legislation routes moved under committees
Legislation must always be for a particular committee.  Because of this, we really wanted to have the legislation URLs be underneath the committee URLs.  This means the committee_id is always provided as part of the path.

All of the pages dealing with legislation routes had to be updated to read the committee_id from the params, rather than from the request.