Report search now shows committee name
The reports listing, like the meeting files were originally written expecting to be used in the context of a committee.  The search form now allows for searching for reports for any and all committees.

If there is no committee in the context, then we must display the committee on each search result.

Updates #397