Declare USWDS url in configuration
We do not want to include the compiled USWDS distribution in our source code.  It is an external dependency that must be hosted on a web server somewhere.

However, because we're relying on SVG files for the Icon component, we are limited to hosting it on the same server as the application.

Updates #384