Committee: termEndWarning used conditionally
The two field, termEndWarningDays and applicationLifetime are no longer included on the edit form when they are not relevant.   When the fields are included on the form, they are marked as required.

The database fields are not null and have a default value.  The default value is meaningless when not needed.  So, it should not hurt to have values in the table even when they are not relevant.

Updates #411