Rename ID parameters and add to REQUEST
It was a mistake to use 'ID' as the parameter name regardless of context.  It caused confusion during access control checks.  This establishes all parameter names needed for the whole system.  We'll need to implement the parameters one route at a time, though.

The controllers also need a consistent way to access parameters.  The ROUTE objects do not contain al the parameters for the request, only those that were regex from the path.  The ROUTE attributes must be added to every REQUEST.