Assign public user accounts by email address
Access Indiana verifies people's email address and uses the email address as the username.  When Access Indiana gives us a username, we're going to trust that anyone in the system with that username or email address is the same person who is behind the browser coming from Access Indiana.

Updates #404