Display all meetings that occur at the same time
It may be the case that boards have multiple simultaneous meetings.  The code has always presumed there would only be one meeting for a committee at a given time.

Since this is not actually true, we need to add an additional level of nesting in Committee::meetings.  This is a breaking change for things that rely on the JSON web service.

Updates #442