Change route for Google push notifications
When this was first created, "notifications" seemed like a good word to use in the URL.  Now that we want to build a notification system, we want to reserver the word "notification" for the new notification system.

The Google push notification route will need to be called something else.  It really doesn't matter what the Google route is called, as it is only configured in one place and only Google calls it.  It is not a URL we will link to from anywhere in the system.

Updates #433