Fix: pagination now expects current uri
This update passes the PSR-7 REQUEST to the Twig environment.

Because route parameters are part of the rendered uri, pagination needs to know the current rendered uri.  Pagination cannot re-render the route without knowing the route parameters.

Macros cannot read from the global context, so, to make this information available to pagination, the current URI path must be passed into pagination as the first function parameter

Updates #467